Resolved docs issues
